+++ src/http.js
@@ -2,7 +2,10 @@ var lunarisExports = require('./exports.js');
 var utils          = require('./utils.js');
 
 var baseOptions = {
-  onComplete : null
+  onComplete : null,
+  headers    : {
+    'Client-Version' : 2
+  }
 };
 
 /**
@@ -34,9 +37,7 @@ function setup (options) {
 function request (method, request, body, callback, options) {
   var _body               = body;
   var _defaultContentType = 'application/json';
-  var _headers            = {
-    'Client-Version' : 2
-  };
+  var _headers            = utils.clone(baseOptions.headers);
 
   options = options || {};
   _headers['Content-Type'] = options['Content-Type'] || _defaultContentType;