+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## 1.xx.x
 *2021-XX-XX*
  - [Builder|Vue template compiler] Do not remove spaces at start and end of a tag.
+ - `lunaris.http.setup` can set default http headers, see below:
+    ```js
+      lunaris.http.setup({
+        headers : {
+          'Accept-Language' : 'fr',
+        }
+      });
+    ```
 
 ## 1.22.5
 *2021-06-23*